requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Assets==2.0
Flask-Scss==0.5
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
pyScss==1.3.7
six==1.15.0
webassets==2.0
Werkzeug==1.0.1
